Reapply "Create QR page"
This reverts commit 6770a6137b1c1d55eb73ad0345a9d98218fee2f8.
QR generate page
