+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --turbopack",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
@@ -13,14 +13,16 @@
     "@prisma/client": "^6.0.1",
     "bcrypt": "^5.1.1",
     "cookies-next": "^5.0.2",
-    "next": "15.0.4",
+    "next": "^15.1.0",
     "next-auth": "^4.24.11",
     "prisma": "^6.0.1",
+    "qrcode": "^1.5.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@types/node": "^20",
+    "@types/qrcode": "^1.5.0",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^8",
