+++ src/app/page.tsx
@@ -1,82 +1,42 @@
-// app/users/page.js
-import { PrismaClient } from "@prisma/client";
+"use client";
+import { type NextPage } from "next";
+import Head from "next/head";
+import { useState } from "react";
+import QRCode from "qrcode";
 
-const prisma = new PrismaClient();
+const Home: NextPage = () => {
+  const [url, setUrl] = useState<string>("");
+  const [src, setSrc] = useState<string>("");
 
-export default async function UsersPage() {
-  const users = await prisma.user.findMany(); 
+  const generate = () => {
+    QRCode.toDataURL(`${url}`).then(setSrc);
+  };
 
   return (
-    <div>
-      <h1>Users</h1>
-      <ul>
-        {users.map((user) => (
-          <li key={user.id}>Name : {user.user_name} email : {user.email} Role : {user.user_role}</li>
-        ))}
-      </ul>
-    </div>
+    <>
+      <Head>
+        <title>Create T3 App</title>
+        <meta name="description" content="Generated by create-t3-app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      {/* The QRCode img */}
+      <img src={src} />
+
+      {/* username input field */}
+      <input
+        className="bg-black text-white"
+        type="text"
+        value={url}
+        onChange={(e) => setUrl(e.target.value)}
+      />
+
+      {/* generate button*/}
+      <button type="button" onClick={generate}>
+        Generate
+      </button>
+    </>
   );
-}
+};
 
-// 'use client';
-
-// import { signIn } from 'next-auth/react';
-// import { useState, FormEvent } from 'react';
-
-// const SignIn: React.FC = () => {
-//   const [email, setEmail] = useState<string>('');
-//   const [password, setPassword] = useState<string>('');
-//   const [error, setError] = useState<string>('');
-//   const [isLoading, setIsLoading] = useState<boolean>(false);
-
-//   const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
-//     e.preventDefault();
-//     setIsLoading(true);
-
-//     const res = await signIn('credentials', {
-//       redirect: false,
-//       email,
-//       password,
-//     });
-
-//     setIsLoading(false);
-
-//     if (res?.error) {
-//       setError(res.error || 'Invalid credentials');
-//     } else {
-//       window.location.href = '/Add-course'; // Redirect on successful login
-//     }
-//   };
-
-//   return (
-//     <div className="bg-white text-black">
-//       <h1>Sign In</h1>
-//       <form onSubmit={handleSubmit}>
-//         <div>
-//           <label>Email</label>
-//           <input
-//             type="email"
-//             value={email}
-//             onChange={(e) => setEmail(e.target.value)}
-//             required
-//           />
-//         </div>
-//         <div>
-//           <label>Password</label>
-//           <input
-//             type="password"
-//             value={password}
-//             onChange={(e) => setPassword(e.target.value)}
-//             required
-//           />
-//         </div>
-//         {error && <p>{error}</p>}
-//         <button type="submit" disabled={isLoading}>
-//           {isLoading ? 'Signing In...' : 'Sign In'}
-//         </button>
-//       </form>
-//     </div>
-//   );
-// };
-
-// export default SignIn;
+export default Home;
